Added permissions check in deleteItem mutation resolver

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -58,9 +58,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     //1. find item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user {id}}`);
     //2. check if they own or have permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem && hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
     //3. delete it
     return await ctx.db.mutation.deleteItem({ where }, info);
   },
